+++ apps/bridge-dapp/CHANGELOG.md
@@ -173,3 +173,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Leaf Index Checking & Relayer State
 - Adopt use-query-params Hook, Eliminate Excessive useEffect for Default State Handling
 - Fix Invalid BigNumber String on Tranfer flow
+
+## [1.0.2] - 2023-09-20
+
+### Fixed, Changed and Added
+
+- Adding SEO for Hubble Bridge
+- New Orbit & Tangle bridge deployment